Main_Cyber_Owl_App/requirements.txt
flask
flask-socketio
flask-cors
python-dotenv
pymongo
certifi
psutil
httpx
nltk
numpy<2
scikit-learn
pyahocorasick
mss
opencv-python<4.10
nudenet
deepface
supabase
gunicorn
flask-socketio
eventlet
setuptools
transformers
torch
langdetect
soundcard
SpeechRecognition
sounddevice
